requirements.txt
python-pptx>=1.0.0
pandas>=1.5.0
openpyxl>=3.0.0  # Needed for column width adjustment